src - Refactor ZoneDB and CBRFCZone class
Turn the ZoneDB class into a base class to query from the DB.
This moves zone specific queries into the CBRFZone class and
extracts the dataclasse into a CBRFCZoneRow factory.
