+++ src/swed_17/zone_db.py
@@ -3,17 +3,14 @@ import psycopg
 from contextlib import contextmanager
 
 from psycopg import Cursor
-from psycopg.rows import TupleRow, class_row
-from rasterio.io import MemoryFile
-
-from swed_17.db_tables.cbrfc_zone import CBRFCZone
+from psycopg.rows import TupleRow
+from sqlalchemy import create_engine
 
 
 class ZoneDB:
     """
-    Database query class.
+    Base database query class.
     """
-    COORDS_CONVERSION = {'x': 'lon', 'y': 'lat'}
 
     CONNECTION_OPTIONS = dict(
         autocommit=True,
@@ -61,59 +58,3 @@ class ZoneDB:
                 cursor.execute(query, params)
                 yield cursor
 
-    def zone_as_rio(self, zone_name: str) -> MemoryFile:
-        """
-        Query for a zone mask and return as a rasterio MemoryFile.
-
-        Parameters
-        ----------
-        zone_name : str
-            SQL query
-
-        Returns
-        -------
-        MemoryFile
-            Result of query as rasterio MemoryFile
-        """
-        with self.query(
-            self.Query.ZONE_AS_RASTER, {'zone_name': zone_name}
-        ) as db_result:
-            result = db_result.fetchone()
-        return MemoryFile(bytes(result[0]))
-
-    def zone_in_ch5_ids(self, ch5_ids: list) -> list[str]:
-        """
-        Return list of zone names for given CH5 IDs
-
-        Parameters
-        ----------
-        ch5_ids : list
-            CH5 IDs
-
-        Returns
-        -------
-        Row
-            List with zone names
-        """
-        with self.query(
-            CBRFCZone.ZONES_IN_CH5ID,
-            [ch5_ids],
-            row_factory=class_row(CBRFCZone)
-        ) as db_result:
-            return db_result.fetchall()
-
-    def zone_in_ch5_id(self, ch5_id: str) -> list[str]:
-        """
-        Return list of zone name for given CH5 ID
-
-        Parameters
-        ----------
-        ch5_id : str
-            CH5 ID
-
-        Returns
-        -------
-        list[str]
-            List with zone name
-        """
-        return self.zone_in_ch5_ids([ch5_id])
